requirements.txt
python-telegram-bot[job-queue]>=20.0
sympy>=1.12
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pint>=0.22
cryptography>=41.0
openai>=0.28
google-generativeai>=0.3
anthropic>=0.18